requirements.txt
pandas
numpy
librosa
openai-whisper
nltk
scikit-learn
matplotlib
seaborn
streamlit
soundfile
